ENG-16252: Catch and log all throwables
A scheduled executor catches all throwables and sets the throwable as a failure
on the associated future. If the future is discarded that means that any
throwable generated by a runnable or a callable will never be seen. Instead of
letting the executor handle the exception catch and log all exceptions caused
by migrate leader.